fix: Updated t/lib/T/Constants.pm
CPANTesters report for Perl 5.28.1 on a BSD system shows Wstat 139 SEG fault in t/00-unit-tests/02.t; I have no way to replicate the issue, but will try installing on a macOS system running Perl 5.28.1.

In the meantime, I'm adding a couple of extra flags to the SQLite database connection attributes that may positively contribute to a better outcome.