fix: bug sometimes causing final buffer of results to be uncollected
The final buffer fetch would be enough to cause the statement to finish, resulting in the $sth->{Active} check to cause the final buffer of results to be left behind. We must also check that the cache is empty if the statement is no longer active.

Reflect the same checks in the `is_active` and `not_active` predicates.